requirements.txt
scikit-learn
scikit-image
pillow
numpy
gensim
transformers
torch
torchvision